Prompt for gis buffers when customization is set
Prompt by default for magik-utils-get-buffer-mode when
magik-utils-by-default-prompt-buffer-p is set.

This allows for a workflow to prompt by default for
gis or class-browser buffers when they are not visible or there are
multiple sessions running.
Without the need for a prefix-argument.

Also when multiple gis sessions are running, or multiple gis sessions/buffers
have been made over the course of the emacs session. Then the DEFAULT value
is wrong most of the time. And because of that the interactive command
which uses magik-utils-get-buffer-mode fails.

This new functionality is put behind a customization option so that
users can choose which behaviour they want. And to not interrupt
people's current workflow.