[yasnippet] Search for the closest class name
The previous approach captured the first match based on the list of patterns, but if e.g. a `def_slotted_exemplar` was closer to the starting-point, it would not be used since it found a `_method` earlier in the file.

```magik
_method a.b
_endmethod

def_slotted_exemplar(:b, {})
$

<point here>
```

Previously, it would find `a` as class-name, but you would expect `b` as the class-name.