Fix for restarting a session after the previous one was killed
Starting a new session (using F2-z) only worked if you had the previous session buffer active.

If you did not have the previous buffer active or if you killed the buffer completely, it could not start.

Added `t` as predicate for magik-utils-get-buffer-mode, because otherwise the predicate would be to only take buffers into account that had an active process.

Added `(generate-new-buffer-name (or magik-session-buffer magik-session-buffer-default-name))` in case of the previous Magik Session buffer was killed completely, but F2-z still knew the command. This would then create a new buffer with the old command.